# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Description",
        "title": "Simple Calculator API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/add": {
            "post": {
                "description": "return sum of two numbers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "addition",
                "parameters": [
                    {
                        "description": "comment",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smpl_calc_api.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/div": {
            "post": {
                "description": "return division of two numbers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "division",
                "parameters": [
                    {
                        "description": "comment",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smpl_calc_api.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/mul": {
            "post": {
                "description": "return multiplication of two numbers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "multiplication",
                "parameters": [
                    {
                        "description": "comment",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smpl_calc_api.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/sub": {
            "post": {
                "description": "return subtraction of two numbers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "subtraction",
                "parameters": [
                    {
                        "description": "comment",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smpl_calc_api.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handlers.Response": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "smpl_calc_api.Data": {
            "type": "object",
            "properties": {
                "operand1": {
                    "$ref": "#/definitions/smpl_calc_api.Operand"
                },
                "operand2": {
                    "$ref": "#/definitions/smpl_calc_api.Operand"
                },
                "toBase": {
                    "type": "integer"
                }
            }
        },
        "smpl_calc_api.Operand": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "integer",
                    "example": 10
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}